Fix trying multiple mo-files if one is not found
If the filter 'fallback_locale' is used and multiple locales
are added to check for the old code would exit if the first
mo-file is not found. Now the foreach loop continues and
tries the next mo-file.

Is a mo-file found the function returns true to exit otherwise
it continues and will return false if no new mo-file is found